Tests/LibWeb/Text/expected/navigate-unsupported-type-downloads.txt

download requested for octet-stream: true
iframe rendered the response: false
